Replace pagination proerties with just pagination
