--- src/utils/convertFiltersIntoSequalizeObject.ts
@@ -1,7 +1,0 @@
-import replaceFilterOperators from "./replaceFilterOperators"
-
-const convertFiltersIntoSequalizeObject = (filters) => {
-  return filters ? replaceFilterOperators(filters) : {}
-}
-
-export default async (filters) => convertFiltersIntoSequalizeObject(filters)
