+++ src/utils/convertFiltersIntoSequelizeObject.ts
@@ -1,0 +1,7 @@
+import replaceFilterOperators from "./replaceFilterOperators"
+
+const convertFiltersIntoSequelizeObject = (filters) => {
+  return filters ? replaceFilterOperators(filters) : {}
+}
+
+export default async (filters) => convertFiltersIntoSequelizeObject(filters)
